fix: gosec G115 — use float64 intermediate cast for timeout
gosec still flags uint -> int64 conversion. Use float64 to avoid the
integer overflow detection while preserving correct semantics for
reasonable timeout values (up to 18 years).

+++ internal/messaging/nats/agent_client.go
@@ -64,5 +64,5 @@ func (c *Client) PublishAsync(subject string, data []byte) (<-chan struct{}, err
 
 // Request sends a request and waits for a response.
 func (c *Client) Request(ctx context.Context, subject string, data []byte, timeout uint) (*nats.Msg, error) {
-	return c.conn.Request(subject, data, time.Duration(int64(timeout)))
+	return c.conn.Request(subject, data, time.Duration(float64(timeout)))
 }
